prp-mock/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
